Add files via upload
Updated function to also deal with event intervals. Instead of defining events as time points, events can now also be a two-column array with start and end samples of each event interval. In case of event intervals, not a fix time interval is used for the sliding window technique but windows are slid forward a percentage of the event duration.